Move step by step "app helpers" into presenters
These classes are no longer used outside of this gem.
